Allow for full path specification
It is likely that this script will want to be called from any
location, and create repositories outside the working directory.

Why not allow a full path to be created, and the final path
folder name be used as the repo name.  This encourages th best practice
to have your working directory match the repository name.